[complex] Доделал анимацию дозатора. Теперь он правильно работает!

+++ meln_logic/src/devices.rs
@@ -19,8 +19,8 @@ impl Dozator {
 //             hz: 0,
         })
     }
-    pub fn set_value(&self, value: f32) {
-        self.direct.set_bit(value>=0.0);
+    pub fn set_value(&self, value: i32) {
+        self.direct.set_bit(value>=0);
         self.hz_value.set_value(value as u32);
     }
     pub fn set_value_stream(&self, finish_value: i32) -> impl Stream<Item = i32>  {
